Fixed a SEGFAULT in setPose reported in #18
Previously, calling setPose() on an interactive marker causes a SEGFAULT
if applyChanges() was not called on the server at least once since the
marker was created. I traced the actual SEGFAULT to the doSetPose
function. The value of header passed from setPose() is invalid because,
in this case, marker_context_it = marker_contexts_.end().

I added a check for this case and, if there is no marker is present,
instead use the header from the pending update.